mong repository with generated id

+++ SpringBootRest/src/main/java/com/sm/samples/warehouse/dao/ItemDao.java
@@ -1,9 +1,11 @@
 package com.sm.samples.warehouse.dao;
 
+import java.math.BigInteger;
+
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 import com.sm.samples.warehouse.model.ItemModel;
 
-public interface ItemDao extends MongoRepository<ItemModel, Long> {
+public interface ItemDao extends MongoRepository<ItemModel, BigInteger> {
 
 }
